backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
qrcode[pil]==7.4.2
Pillow==10.1.0
